Modified the units to adhere to the new standard for units
