+++ src/assets/Flower.java
@@ -1,5 +1,7 @@
 package assets;
 
+import engine.Board;
+
 /**
  * The Flower class is used for initializing a sunflower that provides points to the user
  * to purchase more plants
@@ -8,8 +10,8 @@ package assets;
  */
 
 public class Flower extends Plant{
-	private static final int DEFAULT_HP = 2;
-	private static final int DEFAULT_POWER = 0;
+	private static final int DEFAULT_HP = HEALTH_LOW;
+	private static final int DEFAULT_POWER = ATTACK_NONE;
 	private static final int COST = 25;
 	
 	//The Amount of Points that are added per turn per sunflower
@@ -42,7 +44,7 @@ public class Flower extends Plant{
 	}
 	
 	@Override
-	public void attack(Zombie zombie) {
-		//unimplemented for sunflower
+	public void attack(Board board) {
+		return; //cannot attack
 	}
 }
