+++ src/assets/SprintZombie.java
@@ -1,9 +1,9 @@
 package assets;
 
 public class SprintZombie extends Zombie {
-	private static final int DEFAULT_SPEED = 3;
-	private static final int DEFAULT_POWER = 1;
-	private static final int DEFAULT_HP = 4;
+	private static final int DEFAULT_SPEED = SPEED_HIGH;
+	private static final int DEFAULT_POWER = ATTACK_LOW;
+	private static final int DEFAULT_HP = HEALTH_LOW;
 	private static final ZombieTypes ZOMBIE_TYPE = ZombieTypes.SPRINT_ZOMBIE;
 	
 	public SprintZombie()	{
@@ -14,10 +14,10 @@ public class SprintZombie extends Zombie {
 	 * returns the name of regular type zombie
 	 */
 	public String toString() {
-		return "Cheetah Zombie";	
+		return "Cheetah Zombie";
 	}
 
 	public ZombieTypes getZombieType() {
 		return ZOMBIE_TYPE;
-	} 
+	}
 }
